+++ src/main/cpp/util/utilities.h
@@ -96,7 +96,7 @@ namespace vaultdb {
         template<typename T>
         static bool vectorEquality(const vector<T> & lhs, const vector<T> & rhs)  {
             if(lhs.size() != rhs.size()) return false;
-            for(int i = 0; i < lhs.size(); ++i) {
+            for(size_t i = 0; i < lhs.size(); ++i) {
                 if(lhs[i] != rhs[i]) return false;
             }
             return true;
